Rename block_ids.py to module_ids.py
* rename file
* fix imports and references
