+++ caikit_huggingface_demo/runtime/tasks/text_generation.py
@@ -13,7 +13,7 @@
 # limitations under the License
 
 # Third Party
-from block_ids import TEXT_GENERATION
+from module_ids import TEXT_GENERATION
 from runtime.data_model.results import Text
 from runtime.hf_base import HFBase
 from transformers import AutoModelForCausalLM, AutoTokenizer
